chore: added button validation and real title sending
