+++ lib/repositories/saved_search/saved_search_repository.dart
@@ -1,3 +1,3 @@
 abstract class SavedSearchRepository<T> {
-  Future<bool> postSavedSearch(String userId, T savedSearch);
+  Future<bool> postSavedSearch(String userId, T savedSearch, String title);
 }
